General refactoring.
This commit makes the following changes:

1. Refactor AFL output directory parsing into a separate module.

This change allows future extensibility of the tool to support
other fuzzers other than AFL. More work has to be done to encode
the collector module name into the atriage.db file.

2. Rename collect.Results to db.AtriageDB

The name change better results the purpose of the class which is
to be a general database for the tool.