Change `exploitable` to use `utils.run_killchild`.
The `exploitable` command uses GDB to process the files. GDB
spawns a child process that does not get killed on TimeoutExpired
if the normal `subprocess.run` is used.

This commit introduces a `utils.run_killchild` function that is
exactly the same as `subprocess.run` except that the entire
process group is killed.

The `exploitable` command has been changed to use
`utils.runkillchild` instead of `subprocess.run`.