Make `atriage` work with single level output directories.
The current atriage AFL-fuzz collector expects that the output
directories contain subdirectories for each individual fuzzer
process as like when a fuzz job is launched with afl-launch.

This commit changes the collector so that it works for single
level output directories like those you get when you launch a job
with the normal afl-fuzz command.