Make path relative to the `DB` file passed into atriage.
All the paths stored in `atriage.db` is relative to the location
of that particular DB file when the `triage` command is called.

This commit makes the various commands that are used after the
`triage` command reference the crash case path relative to the
location of the `DB` parameter. This will make calling `atriage
gather` (as an example atriage command) from any $CWD work.